// egasApp2-main/admin-dashboard/src/types/driver.ts
export interface Driver {
  id: string;
  name: string;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  phoneNumber: string;
  address: string;
  dateCreated: string;
  no: number;
  phoneNo: string;
  status: string;
  action: string;
  isActive: boolean;
  isBlocked: boolean;
  isDeleted: boolean;
  isAvailable: boolean;
} 